+++ tests/e2e/Scopes/SideNone.php
@@ -8,4 +8,12 @@ trait SideNone
     {
         return [];
     }
+
+    /**
+     * @return string
+     */
+    public function getSide()
+    {
+        return 'none';
+    }
 }
